gestalt_streamlit_template/pyproject.toml
[project]
name = "gestalt-streamlit-template"
version = "0.1.0"
description = "A simple streamlit base"
authors = [
    {name = "Luciano Bermudez",email = "lucianobmecheng69@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "streamlit (>=1.53.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langgraph-sdk (>=0.3.3,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
